+++ src/main/java/io/github/astrapi69/entity/treeable/TreeEntity.java
@@ -26,8 +26,8 @@ package io.github.astrapi69.entity.treeable;
 
 import java.io.Serializable;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
+import javax.persistence.FetchType;
 import javax.persistence.ForeignKey;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
@@ -39,11 +39,12 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
-import io.github.astrapi69.entity.base.SequenceBaseEntity;
+import io.github.astrapi69.entity.base.AutoBaseEntity;
 
 /**
- * The Entity class {@link TreeableEntity} can keep information with a tree structure. The root
- * {@link TreeableEntity} has no parent, all other {@link TreeableEntity} objects have a parent.
+ * The Entity class {@link TreeEntity} can keep information for a tree structure. The root
+ * {@link TreeEntity} has no parent, all other {@link TreeEntity} objects have a parent. The
+ * generation strategy type of the primary key is auto
  */
 @MappedSuperclass
 @Getter
@@ -51,9 +52,9 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @SuperBuilder
-public abstract class TreeableEntity<PK extends Serializable, T, TR extends Treeable<T, TR>>
+public abstract class TreeEntity<PK extends Serializable, T, TR extends Treeable<T, TR>>
 	extends
-		SequenceBaseEntity<PK>
+		AutoBaseEntity<PK>
 	implements
 		Treeable<T, TR>
 {
@@ -67,8 +68,8 @@ public abstract class TreeableEntity<PK extends Serializable, T, TR extends Tree
 	boolean node;
 
 	/** The parent tree entity that references to the parent. */
-	@ManyToOne(cascade = CascadeType.ALL)
-	@JoinColumn(name = "parent_id", referencedColumnName = "id", foreignKey = @ForeignKey(name = "fk_treeable_parent_id"))
+	@ManyToOne(fetch = FetchType.EAGER)
+	@JoinColumn(name = "parent_id", foreignKey = @ForeignKey(name = "fk_treeable_parent_id"))
 	TR parent;
 
 	/** The value of this tree entity */