+++ src/main/java/io/github/astrapi69/entity/treeable/TreeUUIDEntity.java
@@ -22,59 +22,53 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.entity.base;
+package io.github.astrapi69.entity.treeable;
 
-import java.io.Serializable;
-
-import javax.persistence.Access;
-import javax.persistence.AccessType;
 import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
+import javax.persistence.FetchType;
+import javax.persistence.ForeignKey;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.MappedSuperclass;
 
 import lombok.AccessLevel;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
-import io.github.astrapi69.entity.identifiable.Identifiable;
+import io.github.astrapi69.entity.uniqueable.UUIDEntity;
 
 /**
- * The class {@link UniqueableEntity} holds an UUID as primary key
- *
- * The generation strategy type of the primary key is 'UUID'
- *
- * @param <PK>
- *            the generic type of the UUID technical primary key
+ * The Entity class {@link TreeUUIDEntity} can keep information for a tree structure. The root
+ * {@link TreeUUIDEntity} has no parent, all other {@link TreeUUIDEntity} objects have a parent.
  */
 @MappedSuperclass
-@Access(AccessType.FIELD)
 @Getter
 @Setter
 @NoArgsConstructor
-@EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @SuperBuilder
-public abstract class UniqueableEntity<PK extends Serializable> implements Identifiable<PK>
+public class TreeUUIDEntity<T, TR extends Treeable<T, TR>> extends UUIDEntity
+	implements
+		Treeable<T, TR>
 {
 
-	/** The technical primary key. */
-	@Id
-	@EqualsAndHashCode.Include
-	@GeneratedValue(generator = "UUID")
-	@Column(name = "id", updatable = false, nullable = false)
-	PK id;
+	/** The depth of this node. For the root depth would be 0. */
+	@Column(name = "depth")
+	int depth;
 
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public String toString()
-	{
-		return getClass().getSimpleName() + ": id=" + getId();
-	}
+	/** A flag that indicates if this tree entity is a node or a leaf */
+	@Column(name = "node")
+	boolean node;
+
+	/** The parent tree entity that references to the parent. */
+	@ManyToOne(fetch = FetchType.EAGER)
+	@JoinColumn(name = "parent_id", foreignKey = @ForeignKey(name = "fk_treeable_parent_id"))
+	TR parent;
+
+	/** The value of this tree entity */
+	@Column(name = "value", columnDefinition = "TEXT")
+	T value;
 
 }